[5/?] -- do rewriting by annotating a DF to the UDTF table provider instance, since DF disagrees as to which schema to use when rewriting logical plans
