--- crates/runtime/src/search/rrf/mod.rs
@@ -1,2 +1,0 @@
-pub mod optimizer;
-pub mod udf;
